revert back to the push trigger strategy and ignore changes to the workflows.
Reason:
Turns out we cannot deploy from the target of a merge branch (if we do something like main <- dev). This means that I cannot deploy to Github Pages by accepting a pull request from the dev branch and trigger this workflow automatically as it would fail with an error stating that dev cannot deploy to Github Pages. This might be fixed by updating the Github Pages branch settings. However, much easier approach would be just to ignore pushes to workflow files and trigger the workflows when other files change after a commit or a merge.